--- packages.md
@@ -1,63 +1,0 @@
-Generated by [flutlab.io](https://flutlab.io) on 2021-11-29 08:20:51
-
-
-## pub.dartlang.org hosted packages
-
- - [async 2.8.1](https://pub.dartlang.org/packages/async/versions/2.8.1)
- - [audio_session 0.1.6+1](https://pub.dartlang.org/packages/audio_session/versions/0.1.6+1)
- - [boolean_selector 2.1.0](https://pub.dartlang.org/packages/boolean_selector/versions/2.1.0)
- - [characters 1.1.0](https://pub.dartlang.org/packages/characters/versions/1.1.0)
- - [charcode 1.3.1](https://pub.dartlang.org/packages/charcode/versions/1.3.1)
- - [clock 1.1.0](https://pub.dartlang.org/packages/clock/versions/1.1.0)
- - [collection 1.15.0](https://pub.dartlang.org/packages/collection/versions/1.15.0)
- - [crypto 3.0.1](https://pub.dartlang.org/packages/crypto/versions/3.0.1)
- - [cupertino_icons 1.0.4](https://pub.dartlang.org/packages/cupertino_icons/versions/1.0.4)
- - [fake_async 1.2.0](https://pub.dartlang.org/packages/fake_async/versions/1.2.0)
- - [ffi 1.1.2](https://pub.dartlang.org/packages/ffi/versions/1.1.2)
- - [file 6.1.2](https://pub.dartlang.org/packages/file/versions/6.1.2)
- - [js 0.6.3](https://pub.dartlang.org/packages/js/versions/0.6.3)
- - [just_audio 0.9.18](https://pub.dartlang.org/packages/just_audio/versions/0.9.18)
- - [just_audio_platform_interface 4.0.0](https://pub.dartlang.org/packages/just_audio_platform_interface/versions/4.0.0)
- - [just_audio_web 0.4.2](https://pub.dartlang.org/packages/just_audio_web/versions/0.4.2)
- - [matcher 0.12.10](https://pub.dartlang.org/packages/matcher/versions/0.12.10)
- - [meta 1.7.0](https://pub.dartlang.org/packages/meta/versions/1.7.0)
- - [path 1.8.0](https://pub.dartlang.org/packages/path/versions/1.8.0)
- - [path_provider 2.0.7](https://pub.dartlang.org/packages/path_provider/versions/2.0.7)
- - [path_provider_android 2.0.9](https://pub.dartlang.org/packages/path_provider_android/versions/2.0.9)
- - [path_provider_ios 2.0.7](https://pub.dartlang.org/packages/path_provider_ios/versions/2.0.7)
- - [path_provider_linux 2.1.2](https://pub.dartlang.org/packages/path_provider_linux/versions/2.1.2)
- - [path_provider_macos 2.0.3](https://pub.dartlang.org/packages/path_provider_macos/versions/2.0.3)
- - [path_provider_platform_interface 2.0.1](https://pub.dartlang.org/packages/path_provider_platform_interface/versions/2.0.1)
- - [path_provider_windows 2.0.4](https://pub.dartlang.org/packages/path_provider_windows/versions/2.0.4)
- - [platform 3.0.2](https://pub.dartlang.org/packages/platform/versions/3.0.2)
- - [plugin_platform_interface 2.0.2](https://pub.dartlang.org/packages/plugin_platform_interface/versions/2.0.2)
- - [process 4.2.4](https://pub.dartlang.org/packages/process/versions/4.2.4)
- - [rxdart 0.27.3](https://pub.dartlang.org/packages/rxdart/versions/0.27.3)
- - [source_span 1.8.1](https://pub.dartlang.org/packages/source_span/versions/1.8.1)
- - [stack_trace 1.10.0](https://pub.dartlang.org/packages/stack_trace/versions/1.10.0)
- - [stream_channel 2.1.0](https://pub.dartlang.org/packages/stream_channel/versions/2.1.0)
- - [string_scanner 1.1.0](https://pub.dartlang.org/packages/string_scanner/versions/1.1.0)
- - [term_glyph 1.2.0](https://pub.dartlang.org/packages/term_glyph/versions/1.2.0)
- - [test_api 0.4.2](https://pub.dartlang.org/packages/test_api/versions/0.4.2)
- - [typed_data 1.3.0](https://pub.dartlang.org/packages/typed_data/versions/1.3.0)
- - [url_launcher 6.0.16](https://pub.dartlang.org/packages/url_launcher/versions/6.0.16)
- - [url_launcher_android 6.0.13](https://pub.dartlang.org/packages/url_launcher_android/versions/6.0.13)
- - [url_launcher_ios 6.0.13](https://pub.dartlang.org/packages/url_launcher_ios/versions/6.0.13)
- - [url_launcher_linux 2.0.2](https://pub.dartlang.org/packages/url_launcher_linux/versions/2.0.2)
- - [url_launcher_macos 2.0.2](https://pub.dartlang.org/packages/url_launcher_macos/versions/2.0.2)
- - [url_launcher_platform_interface 2.0.4](https://pub.dartlang.org/packages/url_launcher_platform_interface/versions/2.0.4)
- - [url_launcher_web 2.0.4](https://pub.dartlang.org/packages/url_launcher_web/versions/2.0.4)
- - [url_launcher_windows 2.0.2](https://pub.dartlang.org/packages/url_launcher_windows/versions/2.0.2)
- - [uuid 3.0.5](https://pub.dartlang.org/packages/uuid/versions/3.0.5)
- - [vector_math 2.1.0](https://pub.dartlang.org/packages/vector_math/versions/2.1.0)
- - [win32 2.3.1](https://pub.dartlang.org/packages/win32/versions/2.3.1)
- - [xdg_directories 0.2.0](https://pub.dartlang.org/packages/xdg_directories/versions/0.2.0)
-
-## Other packages
-
- - flutter
- - flutter_test
- - flutter_web_plugins
- - sky_engine
- - orgel (this package)
-
